+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # CHANGELOG
 
 ## UNRELEASED
+
+- Fixed unit tests for various plugins
+
+## v1.1.1: 2023-12-13
+
 - Fixed template text for failed plugin
 - Added plugins missing from `setup.py`
+- GCGI-1229: Fixed missing OncoKB definitions
 
 ## v1.1.0: 2023-12-08